+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - `take_until_either`, `take_until_either1`, `take_until_either_and_consume` and `take_until_either_and_consume1`: they can be replaced with `is_not` (possibly combined with something else)
 - `take_until_and_consume`, `take_until_and_consume1`: they can be replaced with `take_until` combined with `take`
 - `sized_buffer` and `length_bytes!`: they can be replaced with the `length_data` function
+- `begin` and `rest_s` function
 
 ## 4.2.3 - 2019-03-23